refactor: migrate to v2
