+++ lib/v2/liquipedia/maintainer.js
@@ -1,3 +1,4 @@
 module.exports = {
     '/counterstrike/matches/:team': ['CookiePieWw'],
+    '/dota2/matches/:id': ['wzekin'],
 };
