+++ website/docs/routes/game.mdx
@@ -318,23 +318,23 @@ So the route is [`/itch/devlogs/teamterrible/the-baby-in-yellow`](https://rsshub
 
 ### PES Mobile Announcement {#konami-pes-mobile-announcement}
 
-<Route author="HenryQW" example="/konami/pesmobile/en/ios" path="/konami/pesmobile/:lang?/:os?" paramsDesc={['language, obtained from the URL, eg. zh-cn, zh-tw, en', 'operating system，iOS or Android']}/>
+<Route author="HenryQW" example="/konami/pesmobile/en/ios" path="/konami/pesmobile/:lang?/:os?" paramsDesc={['language, obtained from the URL, eg. zh-cn, zh-tw, en', 'operating system, iOS or Android']}/>
 
 ## Liquipedia {#liquipedia}
 
 ### Dota2 战队最近比赛结果 {#liquipedia-dota2-zhan-dui-zui-jin-bi-sai-jie-guo}
 
-<Route author="wzekin" example="/liquipedia/dota2/matches/Team_Aster" path="liquipedia/dota2/matches/:id" paramsDesc={['战队名称，可在url中找到。例如:https://liquipedia.net/dota2/Team_Aster']}/>
+<Route author="wzekin" example="/liquipedia/dota2/matches/Team_Aster" path="/liquipedia/dota2/matches/:id" paramsDesc={['战队名称，可在url中找到。例如:https://liquipedia.net/dota2/Team_Aster']}/>
 
 ### Counter Strike Match Results {#liquipedia-cs-match-results}
 
-<Route author="CookiePieWw" example="/liquipedia/counterstrike/matches/G2_Esports" path="liquipedia/counterstrike/matches/:id" path="liquipedia/counterstrike/matches/:id" paramsDesc={['战队名称，可在url中找到。例如:https://liquipedia.net/counterstrike/G2_Esports']}/>
+<Route author="CookiePieWw" example="/liquipedia/counterstrike/matches/G2_Esports" path="/liquipedia/counterstrike/matches/:id" paramsDesc={['战队名称，可在url中找到。例如:https://liquipedia.net/counterstrike/G2_Esports']}/>
 
 ## Maxjia News {#maxjia-news}
 
 ### Dota 2 {#maxjia-news-dota-2}
 
-<Route author="dearrrfish" example="/maxnews/dota2" path="maxnews/dota2" />
+<Route author="dearrrfish" example="/maxnews/dota2" path="/maxnews/dota2" />
 
 ## Minecraft {#minecraft}
